Move OneDrive tests to new file
