+++ src/internal/connector/graph_connector_test.go
@@ -2,9 +2,6 @@ package connector
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/json"
-	"strings"
 	"testing"
 	"time"
 
@@ -17,7 +14,6 @@ import (
 	"github.com/alcionai/corso/src/internal/connector/discovery/api"
 	"github.com/alcionai/corso/src/internal/connector/graph"
 	"github.com/alcionai/corso/src/internal/connector/mockconnector"
-	"github.com/alcionai/corso/src/internal/connector/onedrive"
 	"github.com/alcionai/corso/src/internal/connector/support"
 	"github.com/alcionai/corso/src/internal/data"
 	"github.com/alcionai/corso/src/internal/tester"
@@ -609,34 +605,9 @@ func runRestoreBackupTestVersion0(
 		test.collectionsLatest)
 }
 
-func getTestMetaJSON(t *testing.T, user string, roles []string) []byte {
-	id := base64.StdEncoding.EncodeToString([]byte(user + strings.Join(roles, "+")))
-	testMeta := onedrive.Metadata{Permissions: []onedrive.UserPermission{
-		{ID: id, Roles: roles, Email: user},
-	}}
-
-	testMetaJSON, err := json.Marshal(testMeta)
-	if err != nil {
-		t.Fatal("unable to marshall test permissions", err)
-	}
-
-	return testMetaJSON
-}
-
 func (suite *GraphConnectorIntegrationSuite) TestRestoreAndBackup() {
 	bodyText := "This email has some text. However, all the text is on the same line."
 	subjectText := "Test message for restore"
-
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	// Get the default drive ID for the test user.
-	driveID := mustGetDefaultDriveID(
-		suite.T(),
-		ctx,
-		suite.connector.Service,
-		suite.user,
-	)
 
 	table := []restoreBackupInfo{
 		{
@@ -854,534 +825,11 @@ func (suite *GraphConnectorIntegrationSuite) TestRestoreAndBackup() {
 		// 		},
 		// 	},
 		// },
-		{
-			name:     "OneDriveMultipleFoldersAndFiles",
-			service:  path.OneDriveService,
-			resource: Users,
-			collections: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("a", 33)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-						{
-							name:      "folder-a" + onedrive.DirMetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "folder-a" + onedrive.DirMetaFileSuffix,
-						},
-						{
-							name:      "b" + onedrive.DirMetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "b" + onedrive.DirMetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"folder-a",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("b", 65)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-						{
-							name:      "b" + onedrive.DirMetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "b" + onedrive.DirMetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"folder-a",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("c", 129)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-						{
-							name:      "folder-a" + onedrive.DirMetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "folder-a" + onedrive.DirMetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"folder-a",
-						"b",
-						"folder-a",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("d", 257)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("e", 257)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-			},
-		},
-		{
-			name:     "OneDriveFoldersAndFilesWithMetadata",
-			service:  path.OneDriveService,
-			resource: Users,
-			collections: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("a", 33)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-						{
-							name:      "b" + onedrive.DirMetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "b" + onedrive.DirMetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("e", 66)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-			},
-		},
 	}
 
 	for _, test := range table {
 		suite.T().Run(test.name, func(t *testing.T) {
 			runRestoreBackupTest(
-				t,
-				suite.acct,
-				test,
-				suite.connector.tenant,
-				[]string{suite.user},
-				control.Options{
-					RestorePermissions: true,
-					ToggleFeatures:     control.Toggles{EnablePermissionsBackup: true},
-				},
-			)
-		})
-	}
-}
-
-func (suite *GraphConnectorIntegrationSuite) TestRestoreAndBackupVersion0() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	// Get the default drive ID for the test user.
-	driveID := mustGetDefaultDriveID(
-		suite.T(),
-		ctx,
-		suite.connector.Service,
-		suite.user,
-	)
-
-	table := []restoreBackupInfoMultiVersion{
-		{
-			name:     "OneDriveMultipleFoldersAndFiles",
-			service:  path.OneDriveService,
-			resource: Users,
-
-			collectionsPrevious: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt",
-							data:      []byte(strings.Repeat("a", 33)),
-							lookupKey: "test-file.txt",
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"folder-a",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt",
-							data:      []byte(strings.Repeat("b", 65)),
-							lookupKey: "test-file.txt",
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"folder-a",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt",
-							data:      []byte(strings.Repeat("c", 129)),
-							lookupKey: "test-file.txt",
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"folder-a",
-						"b",
-						"folder-a",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt",
-							data:      []byte(strings.Repeat("d", 257)),
-							lookupKey: "test-file.txt",
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt",
-							data:      []byte(strings.Repeat("e", 257)),
-							lookupKey: "test-file.txt",
-						},
-					},
-				},
-			},
-
-			collectionsLatest: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("a", 33)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-						{
-							name:      "folder-a" + onedrive.DirMetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "folder-a" + onedrive.DirMetaFileSuffix,
-						},
-						{
-							name:      "b" + onedrive.DirMetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "b" + onedrive.DirMetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"folder-a",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("b", 65)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-						{
-							name:      "b" + onedrive.DirMetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "b" + onedrive.DirMetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"folder-a",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("c", 129)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-						{
-							name:      "folder-a" + onedrive.DirMetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "folder-a" + onedrive.DirMetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"folder-a",
-						"b",
-						"folder-a",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("d", 257)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("e", 257)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, test := range table {
-		suite.T().Run(test.name, func(t *testing.T) {
-			runRestoreBackupTestVersion0(
 				t,
 				suite.acct,
 				test,
@@ -1555,375 +1003,6 @@ func (suite *GraphConnectorIntegrationSuite) TestMultiFolderBackupDifferentNames
 			status := backupGC.AwaitStatus()
 			assert.Equal(t, allItems+skipped, status.ObjectCount, "status.ObjectCount")
 			assert.Equal(t, allItems+skipped, status.Successful, "status.Successful")
-		})
-	}
-}
-
-func (suite *GraphConnectorIntegrationSuite) TestPermissionsRestoreAndBackup() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	// Get the default drive ID for the test user.
-	driveID := mustGetDefaultDriveID(
-		suite.T(),
-		ctx,
-		suite.connector.Service,
-		suite.user,
-	)
-
-	table := []restoreBackupInfo{
-		{
-			name:     "FilePermissionsRestore",
-			service:  path.OneDriveService,
-			resource: Users,
-			collections: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("a", 33)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-			},
-		},
-
-		{
-			name:     "FileInsideFolderPermissionsRestore",
-			service:  path.OneDriveService,
-			resource: Users,
-			collections: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("a", 33)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-						{
-							name:      "b" + onedrive.DirMetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "b" + onedrive.DirMetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("e", 66)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-			},
-		},
-
-		{
-			name:     "FileAndFolderPermissionsResote",
-			service:  path.OneDriveService,
-			resource: Users,
-			collections: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("a", 33)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-						{
-							name:      "b" + onedrive.DirMetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "b" + onedrive.DirMetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("e", 66)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-			},
-		},
-
-		{
-			name:     "FileAndFolderSeparatePermissionsResote",
-			service:  path.OneDriveService,
-			resource: Users,
-			collections: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "b" + onedrive.DirMetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "b" + onedrive.DirMetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("e", 66)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-			},
-		},
-
-		{
-			name:     "FolderAndNoChildPermissionsResote",
-			service:  path.OneDriveService,
-			resource: Users,
-			collections: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "b" + onedrive.DirMetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"read"}),
-							lookupKey: "b" + onedrive.DirMetaFileSuffix,
-						},
-					},
-				},
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-						"b",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("e", 66)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      []byte("{}"),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, test := range table {
-		suite.T().Run(test.name, func(t *testing.T) {
-			runRestoreBackupTest(t,
-				suite.acct,
-				test,
-				suite.connector.tenant,
-				[]string{suite.user},
-				control.Options{
-					RestorePermissions: true,
-					ToggleFeatures:     control.Toggles{EnablePermissionsBackup: true},
-				},
-			)
-		})
-	}
-}
-
-func (suite *GraphConnectorIntegrationSuite) TestPermissionsBackupAndNoRestore() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	// Get the default drive ID for the test user.
-	driveID := mustGetDefaultDriveID(
-		suite.T(),
-		ctx,
-		suite.connector.Service,
-		suite.user,
-	)
-
-	table := []restoreBackupInfo{
-		{
-			name:     "FilePermissionsRestore",
-			service:  path.OneDriveService,
-			resource: Users,
-			collections: []colInfo{
-				{
-					pathElements: []string{
-						"drives",
-						driveID,
-						"root:",
-					},
-					category: path.FilesCategory,
-					items: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.DataFileSuffix,
-							data:      []byte(strings.Repeat("a", 33)),
-							lookupKey: "test-file.txt" + onedrive.DataFileSuffix,
-						},
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-					auxItems: []itemInfo{
-						{
-							name:      "test-file.txt" + onedrive.MetaFileSuffix,
-							data:      getTestMetaJSON(suite.T(), suite.secondaryUser, []string{"write"}),
-							lookupKey: "test-file.txt" + onedrive.MetaFileSuffix,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, test := range table {
-		suite.T().Run(test.name, func(t *testing.T) {
-			runRestoreBackupTest(
-				t,
-				suite.acct,
-				test,
-				suite.connector.tenant,
-				[]string{suite.user},
-				control.Options{
-					RestorePermissions: true,
-					ToggleFeatures:     control.Toggles{EnablePermissionsBackup: true},
-				},
-			)
 		})
 	}
 }
